Manage and provide virtual baremetal
prior to use, the ironic nodes need to be managed and provided.
Also added a stub for booting the node, and a note for where
to find the script that creates virtual baremetal nodes from scratch
in case we need to use that instead of the built in ironic config.