Search for appliance rather than hardcoding IP
Use an arp scan to find where the appliance is on the libvirt network
after being booted by vagrant.

Also move the mgmt iface onto a bridge so we can connect it to the
provisioning network more easily later. This is needed so that booting
nodes can get pxe from provisioning and then grab their image via http
from glance, which is on the mgmt iface.